requirements.txt
fastapi
uvicorn
supabase==2.3.0
pydantic
python-dotenv
